+++ CHANGELOG.md
@@ -76,6 +76,8 @@ and this project adheres to
 - cosmwasm-std: Create `Uint64` analogously to `Uint128` with string
   serialization allowing the use of the full uint64 range in JSON clients that
   use float numbers, such as JavaScript and jq.
+- cosmwasm-std: Create const functions `Uint64::new` and `Uint128::new` to
+  create instances in a const context.
 
 [#692]: https://github.com/CosmWasm/cosmwasm/issues/692
 [#706]: https://github.com/CosmWasm/cosmwasm/pull/706